==> 06_adjust_supply.Rmd <==
---
title: "Supply of new workers to BC"
output:
  html_document:
    code_folding: hide
    css: styles.css
runtime: shiny
---

```{r setup, include=FALSE}
library(vroom)
library(tidyverse)
library(here)
library(janitor)
library(fpp3)
library(ggalluvial)
library(plotly)
library(kableExtra)
library(readxl)
library(data.table)
library(arrow)
library(reactable)
library(DT)
library(conflicted)
conflicts_prefer(plotly::layout)
conflicts_prefer(dplyr::filter)
conflicts_prefer(dplyr::select)
conflicts_prefer(lubridate::year)
#constants-----------------------------
options(scipen = 999)
forecast_horizon <- 11
forecast_years <- 2027:2036 #CHANGE THIS YEARLY (OR MAKE DATA DEPENDENT)
knitr::opts_chunk$set(
  message = FALSE,  # hide messages
  warning = FALSE   # hide warnings
)
#functions-------------------

alluvium_prep <- function(new_entrants, immigrants, interprovincial, noc_digit, names){
  new_entrants_ten <- new_entrants|>
    mutate(noc=str_sub(noc, noc_digit, noc_digit))|>
    group_by(noc)|>
    summarize(`New Entrants`=sum(`New Entrants`))

  immigrants_ten <- immigrants|>
    mutate(noc=str_sub(noc, noc_digit, noc_digit))|>
    group_by(noc)|>
    summarize(`Net International In-Migration`=sum(`Net International In-Migration`))

  interprovincial_ten <- interprovincial|>
    mutate(noc=str_sub(noc, noc_digit, noc_digit))|>
    group_by(noc)|>
    summarize(`Net Interregional In-Migration`=sum(`Net Interregional In-Migration`))

  full_join(new_entrants_ten, immigrants_ten)|>
    full_join(interprovincial_ten)|>
    pivot_longer(cols=c(`Net International In-Migration`, `New Entrants`, `Net Interregional In-Migration`), names_to = "stream")|>
    left_join(names)|>
    mutate(desc=paste(noc, desc, sep=": "), 
           desc=str_trunc(desc, 35))
}
alluvium_plot <- function(tbbl){
  ggplot(data = tbbl,
       aes(axis1 = stream, axis2 = desc, y = value)) +
  geom_alluvium(aes(fill = stream)) +
  geom_stratum() +
  geom_text(stat = "stratum",
            aes(label = after_stat(stratum)),
            size=4)+
  labs(fill="Sources of new workers")+
  theme_void()+
  theme(legend.position = "bottom")
}

# get names for NOC Broad Categories and TEER levels-----------------------
broad_names <- vroom(here("data", 
                        "supply_side", 
                        "noc_descriptions(keep)",
                        "noc_2021_version_1.0_-_classification_structure.csv"))|>
  select(noc=contains("noc"),
         desc=`Class title`,
         level= contains("Hierarchical"))|>
  filter(level=="Broad Category")|>
  select(-level)|>
  mutate(noc=as.character(noc))

teer_names <- read_csv(here("data", 
                            "supply_side", 
                            "noc_descriptions(keep)",
                            "teer_description_short.csv"))|>
  transmute(noc=as.character(teer),
            desc=data)

```


```{r}
#'for new entrants we use the count of 14 years olds. 
fourteen <- read_csv(here("data",
                        "supply_side",
                        "Population_Projections.csv"))|>
  filter(Region.Name=="British Columbia",
         Gender=="T")|>
  mutate(age=14, 
         pop=`14`)|>
  select(ref_date=Year,age, pop)

#table 14100327 provides participation rate for both BC and Canada. 
can_bc_part_rates <- vroom(here("data", "supply_side", "14100327.csv"))

#the participation rate for BC: used for new entrants.
bc_part_rate_historic <- can_bc_part_rates|>
  clean_names()|>
  filter(geo=="British Columbia",
         gender=="Total - Gender",
         age_group=="15 to 64 years",
         labour_force_characteristics=="Participation rate"
         )|>
    mutate(part_rate=value/100,
           ref_date=as.numeric(ref_date))|>
    select(ref_date, part_rate)|>
    as_tsibble(index=ref_date)

bc_part_rate <- bc_part_rate_historic|>
  model(ets=ETS(part_rate~trend()),
        tslm=TSLM(part_rate~trend()))|>
  forecast(h=forecast_horizon)|> 
  as_tibble()|>
  group_by(ref_date)|>
  summarize(part_rate=mean(.mean))|>
  filter(ref_date %in% forecast_years)


#Table 17100015 provides Net-migration for BC. 
inter_prov_data <- vroom(here("data", "supply_side", "17100015.csv"))|>
  clean_names()|>
  filter(geo=="British Columbia",
         gender=="Total - gender",
         migrants=="Net-migration",
         age_group=="15 to 64 years")|>
  mutate(ref_date=as.numeric(str_sub(ref_date, 6, 9)))|> #raw dates are of form 2024/2025 
  select(ref_date, value)|>
  as_tsibble(index=ref_date)

inter_prov_stream <- inter_prov_data|>
  model(ets=ETS(value),
        tslm=TSLM(value~trend())
        )|> 
  forecast(h = forecast_horizon)|>
  as_tibble()|>
  group_by(ref_date)|>
  summarize(interprovincial_people=mean(.mean))|>
  filter(ref_date %in% forecast_years)
 
# Table 14100327 provides the participation rate for Canada: applied to net interprovincial
 
can_part_rate_historic <- can_bc_part_rates|>
  clean_names()|>
  filter(geo=="Canada",
         gender=="Total - Gender",
         age_group=="15 to 64 years",
         labour_force_characteristics=="Participation rate"
         )|>
    mutate(part_rate=value/100,
           ref_date=as.numeric(ref_date))|>
    select(ref_date, part_rate)|>
    as_tsibble(index=ref_date)

can_part_rate <- can_part_rate_historic|>
  model(ets=ETS(part_rate~trend()),
        tslm=TSLM(part_rate~trend()))|>
  forecast(h=forecast_horizon)|>
  as_tibble()|>
  group_by(ref_date)|>
  summarize(part_rate=mean(.mean))|>
  filter(ref_date %in% forecast_years)

#Sources of foreign supply

immigrant_data <- vroom(here("data", "supply_side", "17100014.csv"))|>
  clean_names()|>
  filter(geo=="British Columbia",
         gender=="Total - gender",
         type_of_migrant %in% c("Immigrants", "Net emigration", "Net non-permanent residents"),
         age_group=="15 to 64 years")|>
  mutate(ref_date=as.numeric(str_sub(ref_date, 6,9)))|>
  select(ref_date, type_of_migrant, value)|>
  as_tsibble(index=ref_date, key=type_of_migrant)

immigrant_stream <- immigrant_data|>
  model(ets=ETS(value),
        tslm=TSLM(value~trend())
        )|> 
  forecast(h = forecast_horizon)|>
  as_tibble()|>
  group_by(ref_date, type_of_migrant)|>
  summarize(value=mean(.mean))|>
  pivot_wider(id_cols = ref_date, names_from = type_of_migrant, values_from = value)|>
  clean_names()|>
  mutate(permanent_residents=immigrants-net_emigration)|>
  select(-immigrants, -net_emigration)|>
  pivot_longer(cols=-ref_date)|>
  filter(ref_date %in% forecast_years)

#flow of immigrant people  
immigrant_data <- immigrant_data|>
  as_tibble()|>
  pivot_wider(names_from = type_of_migrant, values_from=value)|>
  clean_names()|>
  mutate(permanent_residents=immigrants-net_emigration)|>
  select(-immigrants, -net_emigration)|>
  pivot_longer(cols=-ref_date)

#immigrant participation rates
immigrant_part_rate <- open_dataset(here("data", "supply_side", "98100446_parquet"))|>
  dplyr::filter(GEO == "British Columbia",
                `Immigrant status and period of immigration (11)`%in% c("2016 to 2021","Non-permanent residents"),
                `Highest certificate, diploma or degree (7)`=="Total - Highest certificate, diploma or degree",
                `Age (15A)`%in% c("15 to 24 years","25 to 64 years"),
                `Gender (3)`=="Total - Gender",
                `Visible minority (15)`=="Total - Visible minority",
                `Statistics (3)`=="Count"
                )|>
  select(contains("status"), contains("age"))|>
  collect()|>
  clean_names()|>
  pivot_longer(cols=-c("immigrant_status_and_period_of_immigration_11", "age_15a"))|>
  rename(immigrant_class=contains("immigrant_status"))|>
  mutate(value=as.numeric(value),
         name=str_remove_all(name, "labour_force_status_8_"),
         name=str_sub(name, end=-3))|>
  group_by(immigrant_class, name)|>
  summarize(value=sum(value))|>
  pivot_wider(id_cols = immigrant_class)|>
  mutate(immigrant_part_rate=(unemployed+employed)/(unemployed+employed+not_in_the_labour_force))|>
    select(immigrant_class, immigrant_part_rate)

immigrant_part_rate$immigrant_class[1] <- "Permanent residents"

pr_part_rate <- immigrant_part_rate$immigrant_part_rate[immigrant_part_rate$immigrant_class=="Permanent residents"]
npr_part_rate <- immigrant_part_rate$immigrant_part_rate[immigrant_part_rate$immigrant_class=="Non-permanent residents"]

entrant_noc_props <- open_dataset(here("data",
                                      "supply_side",
                              "98100593_parquet"))|>
  dplyr::filter(GEO == "British Columbia",
                `Statistics (3)`=="Count",
                `Labour force status (3)`=="Employed",
                `Age (15A)`=="15 to 19 years",
                `Gender (3)`=="Total - Gender"
                )|>
  dplyr::select(noc_desc=`Occupation - Unit group - National Occupational Classification (NOC) 2021 (821A)`,
                count=`Class of worker (7A):Total - Class of worker[1]`)|>
  collect()|>
  filter(str_detect(noc_desc, "\\b\\d{5}\\b"))|>
  mutate(count=as.numeric(count),
         noc=str_sub(noc_desc, 1, 5),
           desc=str_sub(noc_desc, 7),
           prop=count/sum(count, na.rm=TRUE)
    )|>
    select(noc, desc, prop)|>
    arrange(desc(prop))

canada_noc_props <- open_dataset(here("data",
                                      "supply_side",
                              "98100449_parquet"))|>
  filter(GEO == "Canada",
                `Highest certificate, diploma or degree (16)`=="Total - Highest certificate, diploma or degree",
                `Age (15A)`=="Total - Age",
                `Gender (3)`=="Total - Gender")|>
  select(noc_desc=`Occupation - Unit group - National Occupational Classification (NOC) 2021 (821A)`,
         count=`Labour force status (3):Employed[2]`
         )|>
  collect()|>
  filter(str_detect(noc_desc, "\\b\\d{5}\\b"))|>
  mutate(count=as.numeric(count),
         noc=str_sub(noc_desc, 1, 5),
           desc=str_sub(noc_desc, 7),
           prop=count/sum(count, na.rm=TRUE)
    )|>
    select(noc, desc, prop)|>
    arrange(desc(prop))

immigrant_noc_props <- open_dataset(here("data",
                                      "supply_side",
                              "98100316_parquet"))|>
  filter(GEO=="Canada",
         `Admission category and applicant type (35)`=="Total - Admission category and applicant type",
         `Highest certificate, diploma or degree (7)`=="Total - Highest certificate, diploma or degree",
         `Age (8C)`=="Total - Age",
         `Gender (3)`=="Total - Gender")|>
  select(noc_desc=`Occupation - Unit group - National Occupational Classification (NOC) 2021 (821A)`,
         count=`Period of immigration (8A):2016 to 2021[8]`)|>
  collect()|>
  filter(str_detect(noc_desc, "\\b\\d{5}\\b"))|>
  mutate(count=as.numeric(count),
         noc=str_sub(noc_desc, 1, 5),
         desc=str_sub(noc_desc, 7),
         prop=count/sum(count, na.rm=TRUE)
    )|>
    select(noc, desc, prop)|>
    arrange(desc(prop))

demand <- read_rds(here("out","richs_forecast.rds"))|>
  group_by(noc_5, year)|>
  summarize(expansion_demand=sum(expansion_demand, na.rm=TRUE),
            replacement_demand=sum(replacement_demand, na.rm=TRUE))|>
  rename(ref_date=year)|>
  filter(ref_date %in% forecast_years)|>
  mutate(`Job Openings`=expansion_demand+replacement_demand)|>
  select(noc=noc_5, ref_date, `Job Openings`)

stokes_total <- read_excel(here("data",
                                "supply_side", 
                                "stokes_data(replace)", 
                                "supply_demand.xlsx"), skip=3)%>%
  pivot_longer(cols= starts_with("2"), names_to = "year", values_to = "value")%>%
  clean_names()|>
  filter(noc=="#T",
         industry=="All Industries",
         geographic_area=="British Columbia",
         variable %in% c("Job Openings",
                         "New Entrants",
                         "Net International In-Migration",
                         "Net Interregional In-Migration",
                         "Net Other Mobility",
                         "Total Supply Change"),
         year>min(year))%>%
  group_by(name=variable)%>%
  summarise(value=sum(value))|>
  pivot_wider()|>
  mutate(`Decline in unemployment`=`Job Openings`-`Total Supply Change`,
         `Change in labour market participation`=`Net Other Mobility`+`Decline in unemployment`)|>
  pivot_longer(cols=everything(), values_to = "stokes (previous)")
```

## Caveat Comedens

> Econometric theory is like an exquisitely balanced French recipe, spelling out precisely with how many turns to mix the sauce, how many carats of spice to add, and for how many milliseconds to bake the mixture at exactly 474 degrees of temperature. But when the statistical cook turns to raw materials, he finds that hearts of cactus fruit are unavailable, so he substitutes chunks of cantaloupe; where the recipe calls for vermicelli he uses shredded wheat; and he substitutes green garment dye for curry, ping-pong balls for turtle's eggs and, for Chalifougnac vintage 1883, a can of turpentine". 
>
> (Stefan Valavanis)


## Streams of people

There are four streams of people that augment British Columbia's supply of labour: 

1) young British Columbians (Entrants), 

and working age^[15-64 years old]

2) Net inter-provincial migrants,
3) New Permanent Residents,
4) Net Non-Permanent Residents. 

In this note we forecast each of these four streams of people as well as their respective participation rates.  When the stream and participation rate are multiplied, we get a forecast of the supply additions to the labour market.  Only by divine miracle would the sum of these four streams exactly equal job openings: any difference is assumed to be balanced by changes in unemployment and participation in working age British Columbians.  

#### Young British Columbians

For the young, we use demographic projections from <https://bcstats.shinyapps.io/popApp/>.  To avoid the possibility of double counting additions of working age^[15-64 years old] people arriving from other jurisdictions, we base our new entrant projections on the counts of 14 year old British Columbians.  Specifically, we shift the age and year associated with the count of 14 year olds to reflect working age threshold. E.g. suppose that there were 50,000 14 year old British Columbians in 2020.  If you assume that people become working age at 23, we would add 9 to both the age and the year, thus giving us our supply of new entrants in 2029. For the participation rate for new entrants, we use Statistics Canada table [14-10-0327-01](https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=1410032701) and filter for BC. Using the controls to the left, one can choose at what age young people cross the threshold to being "working age", and alter the participation rate forecast.   


```{r, fig.retina=2}
fluidPage(
  fluidRow(
    column(
      width = 3,
      sliderInput("working_age", "Working age", min = 16, max = 30, value = 23, step = 1),
      sliderInput("entrant_part_shift", "Young Participation level", min = .98, max = 1.02, value = 1, step = .001),
      sliderInput("entrant_part_slope", "Young Participation slope",min = .999, max = 1.001, value = 1, step = .0001)
      ),
    column(
      width = 9,
      fluidRow(
        column(4, plotOutput("entrant_flow", height = 300)),
        column(4, plotOutput("entrant_part", height = 300)),
        column(4, reactableOutput("entrant_table", height = 300))
      )
    )
  )
)
```

### Net inter-provincial migrants

For net inter-provincial migrants, we use data from Statistics Canada table [17-10-0015-01](https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=1710001501). We filter geo=="British Columbia", gender=="Total - gender", migrants=="Net-migration" and age_group=="15 to 64 years". 

For the participation rate for inter-provincial migrants, we use Statistics Canada table [14-10-0327-01](https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=1410032701) and filter for geo=="Canada", gender=="Total - Gender", age_group=="15 to 64 years" and labour_force_characteristics=="Participation rate".

For both these series we create an ensemble forecast using ETS and TSLM, which is plotted in blue.


```{r, fig.retina=2}
fluidPage(
  fluidRow(
    column(
      width = 3,
      sliderInput("prov_shift", "Interprovincial level", min = .25, max = 4, value = 1, step = .05),
      sliderInput("prov_slope", "Interprovincial slope",min = .9, max = 1.1, value = 1, step = .001),
      sliderInput("prov_part_shift", "Interprovincial participation level", min = .98, max = 1.02, value = 1, step = .001),
      sliderInput("prov_part_slope", "Interprovincial participation slope",min = .9975, max = 1.0025, value = 1, step = .0001)
      ),
    column(
      width = 9,
      fluidRow(
        column(4, plotOutput("prov_flow", height = 400)),
        column(4, plotOutput("prov_part", height = 400)),
        column(4, reactableOutput("prov_table", height = 400))
      )
    )
  )
)
```

#### Immigrants

For immigrants, we use data from Statistics Canada table [17-10-0014-01](https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=1710001401) where we filter geo=="British Columbia", gender=="Total-gender", type of migrant $\in$ (Immigrants, Net emigration, net non-permanent residents), and age_group=="15 to 64 years".  We create a series *permanent_residents* which is the difference between immigrants and net_emigration. We create an ensemble forecast using ETS and TSLM, plotted in blue below. Note that the immigrants component of permanent_residents is consistent with the levels plan for 2026-2028, but we can not use the levels plan to inform our forecast of **net** non-permanent residents. For the participation rates for the two streams (Permanent and Non-permanent residents) of immigrant labour we use Statistics Canada table [98-10-0446-01](https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=9810044601). This is 2021 census data (not LFS time series), so the participation rates are constant over the forecast horizon.  We filter geo=="British Columbia", Immigrant status and period of immigration $\in$ (2016-2021, Non-permanent residents), highest certificate=="Total - Highest certificate, diploma or degree", age $\in$ (15-24, 25-64), Gender=="Total", Visible minority=="Total", Statistics=="Count". For permanent residents, we only consider those who landed between 2016-2021. Because the working age group is split into young/old age brackets, we need to aggregate before calculating working age (15-64) participation rates. 

```{r}
immigrant_part_rate|>
  mutate(immigrant_part_rate=scales::percent(immigrant_part_rate, accuracy = 1))|>
  kableExtra::kable()|>
  kable_styling(full_width = FALSE)
```

#### Permanent Residents

```{r, fig.retina=2}
fluidPage(
  fluidRow(
    column(
      width = 3,
      sliderInput("pr_shift", "PR level", min = .25, max = 4, value = 1, step = .05),
      sliderInput("pr_slope", "PR slope",min = .9, max = 1.1, value = 1, step = .001),
      sliderInput("pr_part_shift", "PR participation rate", min = .98, max = 1.02, value = 1, step = .001),
      ),
    column(
      width = 9,
      fluidRow(
        column(6, plotOutput("pr_flow", height = 300)),
        column(6, reactableOutput("pr_table", height = 300))
      )
    )
  )
)
```

#### Net Non-Permanent Residents

```{r, fig.retina=2}
fluidPage(
  fluidRow(
    column(
      width = 3,
      sliderInput("npr_shift", "NPR level", min = .25, max = 4, value = 1, step = .05),
      sliderInput("npr_slope", "NPR slope",min = .9, max = 1.1, value = 1, step = .001),
      sliderInput("npr_part_shift", "NPR participation rate", min = .98, max = 1.02, value = 1, step = .001),
      ),
    column(
      width = 9,
      fluidRow(
        column(6, plotOutput("npr_flow", height = 300)),
        column(6, reactableOutput("npr_table", height = 300))
      )
    )
  )
)
```

## Noc Proportions:

In order to "inject" the new supply into the labour market, we need to know where,
historically, these sources of new supply typically land (in terms of occupations.)

For young workers, we make use of Statistics Canada Census table [98-10-0593-01](https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=9810059301), where we apply filters geo=="British Columbia", labour force status=="Employed", age=="15 to 19 years", gender=="Total", class of worker=="Employee".  We then calculate proportions for each of the NOCs, to give us an indication of first occupations for young British Columbians.

```{r}
entrant_noc_props|>
  slice_max(prop, n=10)|>
  ggplot(aes(prop, fct_reorder(desc, prop)))+
  geom_col()+
  labs(title="Top NOCs for new entrants.",
       x="Proportion",
       y=NULL)+
  theme_minimal()
```



For inter-provincial migrants, we make use of Statistics Canada Census table [98-10-0449-01](https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=9810044901), and apply filters geo=="Canada", highest certificate=="Total", age=="Total, gender=="Total". We make use of the "Employed" counts to calculate proportions for each of the NOCs, to give us an indication of first occupations for inter-provincial migrants.

```{r}
canada_noc_props|>
  slice_max(prop, n=10)|>
  ggplot(aes(prop, fct_reorder(desc, prop)))+
  geom_col()+
  labs(title="Top NOCs for interprovincial.",
       x="Proportion",
       y=NULL)+
  theme_minimal()
```


For immigrants, we make use of Statistics Canada Census table [98-10-0316-01](https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=9810031601), and apply filters GEO=="Canada", admission category=="total", highest certificate=="total", age=="total", gender=="total", period of immigration=="2016-2021". We then calculate proportions for each of the NOCs, to give us an indication of first occupations for immigrants.

```{r}
immigrant_noc_props|>
  slice_max(prop, n=10)|>
  ggplot(aes(prop, fct_reorder(desc, prop)))+
  geom_col()+
  labs(title="Top NOCs for Immigrants.",
       x="Proportion",
       y=NULL)+
  theme_minimal()
```



## By TEER


```{r, fig.retina=2}
fluidPage(fluidRow(plotOutput("teer_alluvium", height = 600)))
```

## By Broad Occupational


```{r, fig.retina=2}
fluidPage(fluidRow(plotOutput("broad_alluvium", height = 600)))
```

## Inter-occupational mobility.

From the above alluvium plots it should? be obvious that a disproportionate share of new supply is being "injected" into the labour market in TEER 5 Sales and services occupations, necessitating  inter-occupational mobility.  The first step to derive inter-occupational mobility is to calculate supply_no_mobility, which is the sum of the 3 streams of new supply.  We assume that inter-occupational mobility balances supply_no_mobility and demand to the extent possible.  Aggregate supply and aggregate demand will never be exactly equal, so we normalize demand so that it is exactly equal to supply_no_mobility (in aggregate). Inter-occupational mobility is the difference between normalized demand and supply_no_mobility. Supply is the sum of supply_no_mobility and inter-occupational mobility. Finally, the residual is the difference between demand and supply, assumed to be balanced by changes to unemployment and participation among working age British Columbians.

```{r}
tagList(
  downloadButton("download_full", "Download full table"),
  DTOutput("supply_demand")
)
```

## Compare with Stokes

- Note there are a couple definition mis-matches in the table below.

- **Rich:** Total Supply Change just covers the 3 streams of new workers modeled above, it **does not include** Change in labour market participation:
    
$$ \mbox{Total Supply Change} = \mbox{New Entrants} + \mbox{Net International In-Migration} + \mbox{Net Interregional In-Migration} $$
    
$$ \mbox{Job Openings} = \mbox{Total Supply Change} + \mbox{Change in labour market participation} $$
    
- I.e. any imbalance between the demand and the supply model is balanced by Change in labour market participation, conflating changes in participation and unemployment. 

- **Stokes:** Total Supply Change **does include** Change in labour market participation, but nets out changes in Unemployment.

$$
\begin{aligned}
\text{Total Supply Change} =\;& \text{New Entrants} \\
&+ \text{Net International In-Migration} \\
&+ \text{Net Interregional In-Migration} \\
&+ \text{Change in labour market participation} \\
&- \text{Change in Unemployment}
\end{aligned}
$$

$$ \mbox{Job Openings} = \mbox{Total Supply Change} + \mbox{Change in Unemployment} $$

```{r, fig.retina=2}
fluidPage(fluidRow(reactableOutput("results")))
```

## Reactive Elements... if you are interested.

```{r server}
#the differential we add to age and year of the 14 year olds to shift to working age.
age_diff <- reactive({
  input$working_age-14
})

#shifted 14 year olds
entrant_stream_reactive <- reactive({
  fourteen|>
    mutate(age=age+age_diff(),
           ref_date=ref_date+age_diff()
    )
})

#altered bc part rate
bc_part_rate_reactive <- reactive({
  bc_part_rate|>
    mutate(part_rate=part_rate*input$entrant_part_shift,
           part_rate=part_rate*input$entrant_part_slope^(forecast_years - min(forecast_years))
           )
})

#Modified entrant-participants
entrant_participants_reactive <- reactive({
  entrant_stream_reactive()|>
  inner_join(bc_part_rate_reactive())|>
  mutate(new_entrants=pop*part_rate)|>
  select(ref_date, new_entrants)
})

#Modified inter-provincial people
inter_prov_stream_reactive <- reactive({
  inter_prov_stream|>
    mutate(interprovincial_people=interprovincial_people*input$prov_shift,
           interprovincial_people=interprovincial_people*input$prov_slope^(forecast_years - min(forecast_years))
           )
})

#Modified participation rate for canada
can_part_rate_reactive <- reactive({
  can_part_rate|>
    mutate(part_rate=part_rate*input$prov_part_shift,
           part_rate=part_rate*input$prov_part_slope^(forecast_years - min(forecast_years))
           )
})

#Modified inter-provincial participants
inter_prov_participants_reactive <- reactive({
  inter_prov_stream_reactive()|>
  inner_join(can_part_rate_reactive())|>
  mutate(inter_prov=interprovincial_people*part_rate)|>
  select(ref_date, inter_prov)
})

#Modified permanent residents people
pr_stream_reactive <- reactive({
  immigrant_stream|>
    filter(name=="permanent_residents")|>
    ungroup()|>
    mutate(value=value*input$pr_shift,
           value=value*input$pr_slope^(forecast_years - min(forecast_years))
           )
})

pr_part_rate_reactive <- reactive({
  pr_part_rate*input$pr_part_shift
})

pr_participants_reactive <- reactive({
  pr_stream_reactive()|>
  mutate(perm_res=value*pr_part_rate_reactive())|>
  select(ref_date, perm_res)
})

#Modified non permanent residents people
npr_stream_reactive <- reactive({
  immigrant_stream|>
    filter(name=="net_non_permanent_residents")|>
    ungroup()|>
    mutate(value=value*input$npr_shift,
           value=value*input$npr_slope^(forecast_years - min(forecast_years))
           )
})

npr_part_rate_reactive <- reactive({
  npr_part_rate*input$npr_part_shift
})

npr_participants_reactive <- reactive({
  npr_stream_reactive()|>
  mutate(nnpr=value*npr_part_rate_reactive())|>
  select(ref_date, nnpr)
})

immigrant_participants_reactive <- reactive({
  pr_participants_reactive()|>
    full_join(npr_participants_reactive()) |>
     mutate(immigrant_participants=perm_res+nnpr)|>
     select(ref_date, immigrant_participants)
})

new_entrants_reactive <- reactive({
  crossing(entrant_participants_reactive(), entrant_noc_props)|>
  mutate(`New Entrants`=new_entrants*prop)|>
  select(-prop, -desc, -new_entrants)
})

immigrants_reactive <- reactive({
  crossing(immigrant_participants_reactive(), immigrant_noc_props)|>
  mutate(`Net International In-Migration`=immigrant_participants*prop)|>
  select(-immigrant_participants, -prop, -desc)
})

interprovincial_reactive <- reactive({
  crossing(inter_prov_participants_reactive(), canada_noc_props)|>
  mutate(`Net Interregional In-Migration`=inter_prov*prop)|>
  select(ref_date, noc, desc, `Net Interregional In-Migration`)
})

by_teer_reactive <- reactive({
  alluvium_prep(new_entrants_reactive(), immigrants_reactive(), interprovincial_reactive(), 2, teer_names)
})

by_broad_reactive <- reactive({
  alluvium_prep(new_entrants_reactive(), immigrants_reactive(), interprovincial_reactive(), 1, broad_names)
})

supply_no_mobility_reactive <- reactive({
  full_join(immigrants_reactive(), new_entrants_reactive(), by = join_by(ref_date, noc))|>
  full_join(interprovincial_reactive(), by = join_by(ref_date, noc))|>
  mutate(across(where(is.numeric), ~ replace_na(.x, 0)))|>
  select(ref_date, noc, desc, everything())|>
  mutate(supply_no_mobility=`Net International In-Migration`+`New Entrants`+`Net Interregional In-Migration`)
})

supply_and_demand_reactive <- reactive({
  full_join(supply_no_mobility_reactive(), demand)|>
  mutate(normalized_demand=`Job Openings`*sum(supply_no_mobility, na.rm=TRUE)/sum(`Job Openings`, na.rm=TRUE),
         interoccupation_mobility=normalized_demand-supply_no_mobility,
         `Total Supply Change`=supply_no_mobility+interoccupation_mobility,
         `Change in labour market participation`=`Job Openings`-`Total Supply Change`
         )
})

rich_total <- reactive({
  supply_and_demand_reactive()|>
  summarise(across(-c(ref_date, noc, desc), ~ sum(.x, na.rm = TRUE)))|>
  pivot_longer(cols=everything(), values_to = "rich (current)")
})

joined_reactive <- reactive({
  inner_join(rich_total(), stokes_total)
})


#Plots and tables-----------------------

#Flow of new entrants
output$entrant_flow <- renderPlot(
  ggplot(entrant_stream_reactive(), aes(ref_date, pop))+
     geom_rect(aes(xmin = min(forecast_years), xmax = max(forecast_years), ymin = -Inf, ymax = Inf),
            fill = "grey", colour="grey") +
    geom_line()+
    scale_y_continuous(labels=scales::comma, limits=c(40000,60000))+
    scale_x_continuous(limits=c(max(forecast_years)-20, max(forecast_years)))+
    labs(subtitle=paste0("BC residents turning ",input$working_age),
          caption="source: https://bcstats.shinyapps.io/popApp/",
         x=NULL, 
         y=NULL)
  )

#Participation rate for new entrants
output$entrant_part <- renderPlot(
  ggplot()+
    geom_rect(aes(xmin = min(forecast_years), xmax = max(forecast_years), ymin = -Inf, ymax = Inf),
            fill = "grey", colour="grey") +
    geom_line(data=bc_part_rate_historic, mapping=aes(ref_date, part_rate))+
    geom_line(data=bc_part_rate_reactive(), mapping = aes(ref_date, part_rate), colour="blue")+
    scale_y_continuous(labels=scales::percent)+
    labs(subtitle=paste("Participation Rate BC (15-64)"),
         caption="source: Statistics Canada 14-10-0327-01",
         x=NULL, 
         y=NULL)
  )

#New Entrant Forecast
output$entrant_table <- renderReactable({
  entrant_participants_reactive()|>
    adorn_totals()|>
    reactable(columns = list(new_entrants = colDef(format = colFormat(digits = 0, separators = TRUE))),
              pagination = FALSE,
              compact = TRUE,
              bordered = TRUE,
              highlight = TRUE,
              style = list(fontSize = "10px")
              )
})

#Net inter provincial people
output$prov_flow <- renderPlot(
  ggplot()+
    geom_rect(aes(xmin = min(forecast_years), xmax = max(forecast_years), ymin = -Inf, ymax = Inf),
            fill = "grey", colour="grey") +
    geom_line(data=inter_prov_data, mapping=aes(ref_date, value))+
    geom_line(data=inter_prov_stream_reactive(), mapping=aes(ref_date, interprovincial_people), colour="blue")+
    scale_y_continuous(labels=scales::comma)+
    labs(subtitle="Net Inter provincial migrants",
          caption="source: Statistics Canada 17-10-0015-01",
         x=NULL, 
         y=NULL)
  )

# Canada's participation rate
output$prov_part <- renderPlot(
  ggplot()+
    geom_rect(aes(xmin = min(forecast_years), xmax = max(forecast_years), ymin = -Inf, ymax = Inf),
            fill = "grey", colour="grey") +
    geom_line(data=can_part_rate_historic, mapping=aes(ref_date, part_rate))+
    geom_line(data=can_part_rate_reactive(), mapping = aes(ref_date, part_rate), colour="blue")+
    scale_y_continuous(labels=scales::percent)+
    labs(subtitle=paste("Participation Rate Can (15-64)"),
         caption="source: Statistics Canada 14-10-0327-01",
         x=NULL, 
         y=NULL)
  )

#Interprovincial forecast
output$prov_table <- renderReactable({
  inter_prov_participants_reactive()|>
    adorn_totals()|>
    reactable(columns = list(inter_prov = colDef(format = colFormat(digits = 0, separators = TRUE))),
              pagination = FALSE,
              compact = TRUE,
              bordered = TRUE,
              highlight = TRUE
              )
})

#Permanent Resident flow

output$pr_flow <- renderPlot(
  ggplot()+
    geom_rect(aes(xmin = min(forecast_years), xmax = max(forecast_years), ymin = -Inf, ymax = Inf),
            fill = "grey", colour="grey") +
    geom_line(data=immigrant_data|>filter(name=="permanent_residents"), mapping=aes(ref_date, value))+
    geom_line(data=pr_stream_reactive(), mapping=aes(ref_date, value), colour="blue")+
    scale_y_continuous(labels=scales::comma)+
    labs(subtitle="New Permanant Residents",
          caption="source: Statistics Canada 17-10-0014-01",
         x=NULL,
         y=NULL)
  )

#Permanent Resident forecast
output$pr_table <- renderReactable({
  pr_participants_reactive()|>
    adorn_totals()|>
    reactable(columns = list(perm_res = colDef(format = colFormat(digits = 0, separators = TRUE))),
              pagination = FALSE,
              compact = TRUE,
              bordered = TRUE,
              highlight = TRUE,
              style = list(fontSize = "10px")
              )
})

#Non Permanent Resident flow

output$npr_flow <- renderPlot(
  ggplot()+
    geom_rect(aes(xmin = min(forecast_years), xmax = max(forecast_years), ymin = -Inf, ymax = Inf),
            fill = "grey", colour="grey") +
    geom_line(data=immigrant_data|>filter(name=="net_non_permanent_residents"), mapping=aes(ref_date, value))+
    geom_line(data=npr_stream_reactive(), mapping=aes(ref_date, value), colour="blue")+
    scale_y_continuous(labels=scales::comma)+
    labs(subtitle="Net Non Permanant Residents",
          caption="source: Statistics Canada 17-10-0014-01",
         x=NULL,
         y=NULL)
  )

#Net Non Permanent Resident forecast
output$npr_table <- renderReactable({
  npr_participants_reactive()|>
    adorn_totals()|>
    reactable(columns = list(nnpr = colDef(format = colFormat(digits = 0, separators = TRUE))),
              pagination = FALSE,
              compact = TRUE,
              bordered = TRUE,
              highlight = TRUE,
              style = list(fontSize = "10px")
              )
})

output$teer_alluvium <- renderPlot(alluvium_plot(by_teer_reactive()))
output$broad_alluvium <- renderPlot(alluvium_plot(by_broad_reactive()))

output$supply_demand <- renderDT({

  df <- supply_and_demand_reactive()|>
    mutate(ref_date=as.character(ref_date))|>
    arrange(noc, ref_date)
 
  no_underscores <- colnames(df)|>
    str_replace_all("_"," ")
  colnames(df) <- no_underscores
  
  num_cols <- names(df)[vapply(df, is.numeric, logical(1))]

  DT::datatable(
    df,
    filter = "top",
    rownames = FALSE,
    class = "compact stripe hover small-text",
    options = list(
      pageLength = 10,
      paging = TRUE,
      searching = TRUE,
      ordering = TRUE,
      scrollX = TRUE,
      dom = "tip"
    )
  ) |>
    DT::formatRound(columns = num_cols, digits = 2)

}, server = TRUE)

output$download_full <- downloadHandler(
  filename = function() {
    paste0("supply_and_demand_", Sys.Date(), ".csv")
  },
  content = function(file) {
    readr::write_csv(supply_and_demand_reactive(), file)
  }
)

output$results <- renderReactable({
  joined_reactive()|>
      reactable(columns = list(`rich (current)`= colDef(format = colFormat(digits = 0, separators = TRUE)),
                               `stokes (previous)`= colDef(format = colFormat(digits = 0, separators = TRUE)) 
                               ),
               pagination = FALSE,
               compact = TRUE,
               bordered = TRUE,
               highlight = TRUE,
               style = list(fontSize = "10px")
               )
})

```
